+++ internal/service/handler.go
@@ -18,6 +18,9 @@ package service
 
 import "github.com/gorilla/mux"
 
+// APIPrefix is the prefix used in the console-dot environment
+const APIPrefix = "/api/gathering"
+
 // Handler structure represents HTTP request handler.
 type Handler struct {
 	svc Interface
@@ -32,5 +35,6 @@ func NewHandler(svc Interface) *Handler {
 
 // Register function registers new handler for given endpoint URL.
 func (s *Handler) Register(r *mux.Router) {
-	r.Handle("/gathering_rules", gatheringRulesEndpoint(s.svc))
+	r.HandleFunc(APIPrefix+"/openapi.json", serveOpenAPI)
+	r.Handle(APIPrefix+"/gathering_rules", gatheringRulesEndpoint(s.svc))
 }
